created a parsers folder as an intermediary solution.
