+++ parsers/__init__.py
@@ -1,0 +1,12 @@
+__all__ = ["sans",
+           "autoshun",
+           "haleys",
+           "alienvault",
+           "generic",
+           "sans",
+           "rulez",
+           "malwaregroup",
+           "project_honeypot",
+           "packetmail",
+           "virbl",
+           "drg"]
