+++ parsers/alienvault.py
@@ -1,0 +1,15 @@
+from generic import indicator_type
+import datetime
+
+def process(response, source, direction):
+    data = []
+    for line in response.splitlines():
+        if not line.startswith('#') and len(line) > 0:
+            i = line.partition('#')[0].strip()
+            note = line.split('#')[3].strip()
+            if 'Scanning Host' in note or 'Spamming' in note:
+                direction = 'inbound'
+            elif 'Malware' in note or 'C&C' in note or 'APT' in note:
+                direction = 'outbound'
+            data.append((i, indicator_type(i), direction, source, note, '%s' % datetime.date.today()))
+    return data
