+++ parsers/haleys.py
@@ -1,0 +1,10 @@
+from generic import indicator_type
+import datetime
+
+def process(response, source, direction):
+    data = []
+    for line in response.splitlines():
+        if not line.startswith('#') and len(line) > 0:
+            i = line.partition(':')[2].strip()
+            data.append((i, indicator_type(i), direction, source, '', '%s' % datetime.date.today()))
+    return data
