+++ parsers/malwaregroup.py
@@ -1,0 +1,12 @@
+from generic import indicator_type
+import bs4
+
+def process(response, source, direction):
+    data = []
+    soup = bs4.BeautifulSoup(response)
+    for row in soup.find_all('tr'):
+        if row.td:
+            i = row.td.text
+            date = row.contents[-1].text
+            data.append((i, indicator_type(i), direction, source, '', date))
+    return data
