+++ parsers/project_honeypot.py
@@ -1,0 +1,10 @@
+import feedparser
+from generic import indicator_type
+
+def process(response, source, direction):
+    data = []
+    for entry in feedparser.parse(response).entries:
+        i = entry.title.partition(' ')[0]
+        i_date = entry.description.split(' ')[-1]
+        data.append((i, indicator_type(i), direction, source, '', i_date))
+    return data
