+++ parsers/rulez.py
@@ -1,0 +1,10 @@
+from generic import indicator_type
+
+def process(response, source, direction):
+    data = []
+    for line in response.splitlines():
+        if not line.startswith('#') and len(line) > 0:
+            i = line.partition('#')[0].strip()
+            date = line.partition('#')[2].split(' ')[1]
+            data.append((i, indicator_type(i), direction, source, '', date))
+    return data
