+++ parsers/sans.py
@@ -1,0 +1,11 @@
+from generic import indicator_type
+
+def process(response, source, direction):
+    data = []
+    for line in response.splitlines():
+        if not line.startswith('#') and len(line) > 0:
+            # Because SANS zero-pads their addresses
+            i = re.sub('\.0{1,2}', '.', line.split()[0].lstrip('0'))
+            date = line.split()[-1]
+            data.append((i, indicator_type(i), direction, source, '', date))
+    return data
